Add easing and overshoot
The final dial, which moves on its own, was a little bland.  This offers
some variation to its speed as well as a little "overshoot" motion so
that it feels more like a compass wavering around the item.

There is probably still some work to be done around nailing the easing
functions, right now it is using a cubic.  You can see a list of easing
functions over here:

https://gist.github.com/gre/1650294

as well as how they actually look over here:

https://easings.net/

The bounce isn't quite done using easing, but rather through
overshooting and wandering back recursively.

This commit also redoes the way that directionality was determined.  By
using -1 and 1 instead of 0 and 1 it becomes possible remove some
repeated code (since we were either adding or subtracting, now we can
just multiply by the direction in those cases).

Issue #27 have the dials overshoot and snap back to chosen emoji